requirements.txt
requests=2.25.0
Pillow==7.2.0
streamlit==0.87.0
beautifulsoup4==4.9.1
